Finished Set 2, Challenge 12.
Added the 'breakME()' function as decribed in the challenge.
The 'main()' function contains the code that breaks ECB-style encryption
and prints out the decrypted data, one byte at a time.

I could probably optimize the code in the 'main()' function,
but I don't feel it's necessary right now. Maybe later, if I get the itch.